Separate logic from interface, add replay, add model board
