+++ README.md
@@ -25,9 +25,9 @@ _git clone <https://github.com/eypsrcnuygr/Tic-Tac-Toe.git>_
 
 **Step 2**: Direct a terminal into the cloned repository directory
 
-\_cd Tic-Tac-Toe
+_cd Tic-Tac-Toe_
 
-**Step 3**: Run the program from command line, open it in irb, or edit it in a text editor
+**Step 3**: type `main.rb` into the command line, open it in irb, or edit it in a text editor
 
 ## Prerequisities
